Consolidate components where possible
