--- src/components/Portfolio/ExchangeBalance/Deposit/Deposit.module.css
@@ -1,85 +1,0 @@
-.deposit_container {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    padding: 1rem;
-}
-.info_text_non_clickable {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text2);
-}
-.info_text_clickable {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text2);
-    cursor: pointer;
-}
-
-.gas_pump {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: flex-end;
-
-    color: var(--text2);
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-}
-
-.available_container {
-    display: flex;
-    flex-direction: row;
-}
-
-.available_text {
-    margin-right: 6px;
-    color: var(--text1);
-}
-
-.svg_container {
-    background: var(--dark1);
-    width: 20px;
-    height: 20px;
-    padding-right: 6px;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.additional_info {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.max_button {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text1);
-    outline: none;
-    border: none;
-
-    color: var(--text1);
-
-    width: 60px;
-    display: flex;
-    justify-content: left;
-    align-items: center;
-    margin-left: 6px;
-
-    max-height: 20px;
-    transition: all var(--animation-speed) ease-in-out;
-    background: transparent;
-    border-radius: 4px;
-}
-
-.max_button_enabled {
-    cursor: pointer;
-}
-
-.max_button_enabled:hover {
-    color: var(--accent1);
-}
