--- src/components/Portfolio/ExchangeBalance/Deposit/DepositButton/DepositButton.module.css
@@ -1,3 +1,0 @@
-.button_container_disabled button {
-    color: var(--text3);
-}
