+++ src/components/Portfolio/ExchangeBalance/Transfer/Transfer.tsx
@@ -1,11 +1,8 @@
 import { toDisplayQty } from '@crocswap-libs/sdk';
 import { TokenIF } from '../../../../utils/interfaces/exports';
-import styles from './Transfer.module.css';
-import TransferAddressInput from './TransferAddressInput/TransferAddressInput';
-import TransferButton from './TransferButton/TransferButton';
-import TransferCurrencySelector from './TransferCurrencySelector/TransferCurrencySelector';
+import Button from '../../../Global/Button/Button';
+import TransferAddressInput from './TransferAddressInput';
 // import { defaultTokens } from '../../../../utils/data/defaultTokens';
-import { useAppDispatch } from '../../../../utils/hooks/reduxToolkit';
 import {
     Dispatch,
     SetStateAction,
@@ -14,7 +11,27 @@ import {
     useMemo,
     useState,
 } from 'react';
+import { useAppDispatch } from '../../../../utils/hooks/reduxToolkit';
 // import { setToken } from '../../../../utils/state/temp';
+import { BigNumber } from 'ethers';
+import { FaGasPump } from 'react-icons/fa';
+import { getFormattedNumber } from '../../../../App/functions/getFormattedNumber';
+import useDebounce from '../../../../App/hooks/useDebounce';
+import { IS_LOCAL_ENV, ZERO_ADDRESS } from '../../../../constants';
+import { ChainDataContext } from '../../../../contexts/ChainDataContext';
+import { CrocEnvContext } from '../../../../contexts/CrocEnvContext';
+import { FlexContainer, Text } from '../../../../styled/Common';
+import {
+    GasPump,
+    SVGContainer,
+    MaxButton,
+} from '../../../../styled/Components/Portfolio';
+import {
+    TransactionError,
+    isTransactionFailedError,
+    isTransactionReplacedError,
+} from '../../../../utils/TransactionError';
+import { checkBlacklist } from '../../../../utils/data/blacklist';
 import {
     addPendingTx,
     addReceipt,
@@ -22,19 +39,7 @@ import {
     removePendingTx,
     updateTransactionHash,
 } from '../../../../utils/state/receiptDataSlice';
-import {
-    isTransactionFailedError,
-    isTransactionReplacedError,
-    TransactionError,
-} from '../../../../utils/TransactionError';
-import { BigNumber } from 'ethers';
-import { checkBlacklist } from '../../../../utils/data/blacklist';
-import { FaGasPump } from 'react-icons/fa';
-import { IS_LOCAL_ENV, ZERO_ADDRESS } from '../../../../constants';
-import useDebounce from '../../../../App/hooks/useDebounce';
-import { CrocEnvContext } from '../../../../contexts/CrocEnvContext';
-import { ChainDataContext } from '../../../../contexts/ChainDataContext';
-import { getFormattedNumber } from '../../../../App/functions/getFormattedNumber';
+import CurrencySelector from '../CurrencySelector';
 
 interface propsIF {
     selectedToken: TokenIF;
@@ -249,17 +254,16 @@ export default function Transfer(props: propsIF) {
     const isResolvedAddressDifferent = resolvedAddress !== sendToAddress;
 
     const resolvedAddressOrNull = isResolvedAddressDifferent ? (
-        <div className={styles.info_text_non_clickable}>
+        <Text fontSize='body' color='text2'>
             Resolved Destination Address:
-            <div className={styles.hex_address}>{resolvedAddress}</div>
-        </div>
+            <p style={{ userSelect: 'all' }}>{resolvedAddress}</p>
+        </Text>
     ) : null;
 
     const secondaryEnsOrNull = secondaryEnsName ? (
-        <div className={styles.info_text_non_clickable}>
+        <Text fontSize='body' color='text2'>
             Destination ENS Address: {secondaryEnsName}
-            {/* <div className={styles.hex_address}>{secondaryEnsName}</div> */}
-        </div>
+        </Text>
     ) : null;
 
     const resetTransferQty = () => {
@@ -271,18 +275,7 @@ export default function Transfer(props: propsIF) {
         resetTransferQty();
     }, [selectedToken.address]);
 
-    const isTokenDexBalanceGreaterThanZero = parseFloat(tokenDexBalance) > 0;
-
     const [inputValue, setInputValue] = useState('');
-
-    const handleBalanceClick = () => {
-        if (isTokenDexBalanceGreaterThanZero) {
-            setTransferQtyNonDisplay(tokenDexBalance);
-
-            if (tokenExchangeDepositsDisplay)
-                setInputValue(tokenExchangeDepositsDisplay);
-        }
-    };
 
     const [transferGasPriceinDollars, setTransferGasPriceinDollars] = useState<
         string | undefined
@@ -315,58 +308,51 @@ export default function Transfer(props: propsIF) {
     }, [gasPriceInGwei, ethMainnetUsdPrice, isTokenEth]);
 
     return (
-        <div className={styles.deposit_container}>
-            <div className={styles.info_text_non_clickable}>
+        <FlexContainer flexDirection='column' gap={16} padding={'16px'}>
+            <Text fontSize='body' color='text2'>
                 Transfer deposited collateral to another deposit account:
-            </div>
+            </Text>
             <TransferAddressInput
                 fieldId='exchange-balance-transfer-address'
                 setTransferToAddress={setSendToAddress}
                 sendToAddress={sendToAddress}
                 disable={isAddressFieldDisabled}
             />
-            <TransferCurrencySelector
+            <CurrencySelector
                 selectedToken={selectedToken}
-                setTransferQty={setTransferQtyNonDisplay}
+                setQty={setTransferQtyNonDisplay}
                 inputValue={inputValue}
                 setInputValue={setInputValue}
                 disable={isCurrencyFieldDisabled}
                 setTokenModalOpen={setTokenModalOpen}
             />
-            <div className={styles.additional_info}>
-                <div
-                    className={`${styles.available_container} ${styles.info_text_non_clickable}`}
-                >
-                    <div className={styles.available_text}>Available:</div>
+            <FlexContainer justifyContent='space-between' alignItems='center'>
+                <FlexContainer fontSize='body' color='text2'>
+                    <Text margin='0 6px 0 0' color='text1'>
+                        Available:
+                    </Text>
                     {tokenDexBalanceTruncated || '...'}
-                    <button
-                        className={`${styles.max_button} ${
-                            tokenDexBalance !== '0' && styles.max_button_enabled
-                        }`}
-                        onClick={handleBalanceClick}
-                        disabled={tokenDexBalance === '0'}
-                    >
+                    <MaxButton disabled={tokenDexBalance === '0'}>
                         Max
-                    </button>
-                </div>
-                <div className={styles.gas_pump}>
-                    <div className={styles.svg_container}>
+                    </MaxButton>
+                </FlexContainer>
+                <GasPump>
+                    <SVGContainer>
                         <FaGasPump size={12} />{' '}
-                    </div>
+                    </SVGContainer>
                     {transferGasPriceinDollars
                         ? transferGasPriceinDollars
                         : '…'}
-                </div>
-            </div>
+                </GasPump>
+            </FlexContainer>
             {resolvedAddressOrNull}
             {secondaryEnsOrNull}
-            <TransferButton
-                onClick={() => {
-                    transferFn();
-                }}
+            <Button
+                title={buttonMessage}
+                action={transferFn}
                 disabled={isButtonDisabled}
-                buttonMessage={buttonMessage}
+                flat={true}
             />
-        </div>
+        </FlexContainer>
     );
 }
