--- src/components/Portfolio/ExchangeBalance/Transfer/TransferAddressInput/TransferAddressInput.module.css
@@ -1,70 +1,0 @@
-/* Input */
-
-.currency_quantity {
-    font-weight: 300;
-    font-size: 18px;
-    line-height: 22px;
-    color: white;
-    text-align: start;
-    width: 100%;
-    height: calc(1em + 0.65rem + 2px);
-    padding: 0.375rem 0.75rem;
-    border: none;
-    border-radius: 0.25rem;
-    outline: 0;
-    background-color: transparent;
-    background-clip: padding-box;
-
-    transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
-}
-.currency_quantity::placeholder {
-    color: var(--text2);
-    font-weight: 300;
-    font-size: 18px;
-    line-height: 22px;
-}
-
-/* swapbox */
-.swapbox {
-    display: flex;
-    overflow: auto;
-    flex-direction: column;
-    overflow-y: hidden;
-    color: white;
-}
-
-.direction {
-    font-weight: 300;
-    margin-bottom: 0.5rem;
-}
-.direction {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text1);
-}
-
-.swapbox_top {
-    border-radius: var(--border-radius);
-    background-color: var(--dark2);
-    display: flex;
-    justify-content: space-between;
-    flex-direction: row;
-    align-items: center;
-    /* padding: 0.5rem 0; */
-    min-height: 40px;
-    /* height: 40px; */
-}
-
-.swap_input {
-    width: 100%;
-}
-
-.address_display {
-    overflow: hidden;
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text2);
-    padding: 4px;
-    margin: 0 8px;
-    font-family: 'mono-space';
-}
