--- src/components/Portfolio/ExchangeBalance/Transfer/TransferButton/TransferButton.module.css
@@ -1,3 +1,0 @@
-.button_container_disabled button {
-    color: var(--text3);
-}
