--- src/components/Portfolio/ExchangeBalance/Withdraw/WithdrawCurrencySelector/WithdrawCurrencySelector.module.css
@@ -1,84 +1,0 @@
-/* Input */
-
-.currency_quantity {
-    font-weight: 300;
-    font-size: 18px;
-    line-height: 22px;
-    color: white;
-    text-align: start;
-    width: 100%;
-    height: calc(1.5em + 0.75rem + 2px);
-    padding: 0.375rem 0.75rem;
-    border: none;
-    border-radius: 0.25rem;
-    outline: 0;
-    background-color: transparent;
-    background-clip: padding-box;
-
-    transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
-}
-.currency_quantity::placeholder {
-    color: var(--text2);
-    font-weight: 300;
-    font-size: 18px;
-    line-height: 22px;
-}
-
-/* swapbox */
-.swapbox {
-    display: flex;
-    overflow: auto;
-    flex-direction: column;
-    overflow-y: hidden;
-    color: white;
-}
-
-.modalSwapbox {
-    background-color: var(--dark2);
-}
-
-.direction {
-    font-weight: 300;
-
-    padding: 0.4.5px 0;
-    margin: 4px 0;
-}
-.direction,
-.swapbox_bottom {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text1);
-}
-
-.swapbox_top {
-    border-radius: var(--border-radius);
-    background-color: var(--dark2);
-    display: flex;
-    justify-content: space-between;
-    flex-direction: row;
-
-    min-height: 40px;
-}
-.swap_input {
-    width: 70%;
-}
-.token_select {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 4px;
-    cursor: pointer;
-    padding: 0 4px;
-    border-radius: 4px;
-    transition: all var(--animation-speed) ease-in-out;
-}
-.token_select:hover {
-    background: var(--dark3);
-}
-
-.token_list_text {
-    font-size: var(--header2-size);
-    line-height: var(--header2-lh);
-    color: var(--text1);
-    font-weight: 300;
-}
