--- src/pages/Portfolio/PORTFOLIO.md
@@ -1,10 +1,0 @@
-# Introduction
-
-This directory contains the JSX to render a page at the `index/portfolio` URL pathway and all supporting files for that file.
-
-# Location and Files
-
--   This directory is located at `/src/pages/Porfoltio`
--   `Portfolio.tsx` The main file housing the `Portfolio()` react function
--   `Portfolio.module.css` CSS module with styling for `Portfolio.tsx`
--   `PORTFOLIO.md` documentation for all files in this directory
